requirements_core.txt
# Core Framework
fastapi==0.128.0
uvicorn==0.40.0

# AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.21.0
pandas>=2.0.0

# Vector Database & Retrieval
faiss-cpu>=1.7.0
# For GPU support, use: faiss-gpu>=1.7.0

# Web Interface
gradio>=4.0.0

# Model Inference
lmdeploy>=0.11.0

# Utilities
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# Data Processing
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional Dependencies (uncomment if needed)
# redis>=4.5.0  # For Redis memory backend
# accelerate>=0.20.0  # For model acceleration
# bitsandbytes>=0.41.0  # For quantization